Create separate test cases
