+++ test/StreamFactoryTest.php
@@ -2,57 +2,15 @@
 
 namespace Interop\Http\Factory;
 
-use Interop\Http\Factory\StreamFactoryInterface;
-use PHPUnit_Framework_TestCase as TestCase;
-use Psr\Http\Message\StreamInterface;
-
-class StreamFactoryTest extends TestCase
+final class StreamFactoryTest extends StreamFactoryTestCase
 {
-    use StreamHelper;
-
-    /** @var  StreamFactoryInterface */
-    private $factory;
-
-    public function setUp()
+    /**
+     * {@inheritdoc}
+     */
+    protected function createStreamFactory()
     {
         $factoryClass = STREAM_FACTORY;
-        $this->factory = new $factoryClass();
-    }
 
-    private function assertStream($stream, $content)
-    {
-        $this->assertInstanceOf(StreamInterface::class, $stream);
-        $this->assertSame($content, (string) $stream);
-    }
-
-    public function testCreateStream()
-    {
-        $string = 'would you like some crumpets?';
-
-        $stream = $this->factory->createStream($string);
-
-        $this->assertStream($stream, $string);
-    }
-
-    public function testCreateStreamFromFile()
-    {
-        $string = 'would you like some crumpets?';
-        $filename = $this->createTemporaryFile();
-
-        file_put_contents($filename, $string);
-
-        $stream = $this->factory->createStreamFromFile($filename);
-
-        $this->assertStream($stream, $string);
-    }
-
-    public function testCreateStreamFromResource()
-    {
-        $string = 'would you like some crumpets?';
-        $resource = $this->createTemporaryResource($string);
-
-        $stream = $this->factory->createStreamFromResource($resource);
-
-        $this->assertStream($stream, $string);
+        return new $factoryClass();
     }
 }
